--- a/src/fastertransformer/kernels/layernorm_kernels.cu
+++ b/src/fastertransformer/kernels/layernorm_kernels.cu
@@ -1490,6 +1490,17 @@ template void invokeGeneralAddResidualT5PreLayerNorm(half*        output,
                                                      int          n,
                                                      cudaStream_t stream);
 
+#ifdef ENABLE_BF16
+template void invokeGeneralAddResidualT5PreLayerNorm(__nv_bfloat16*       output,
+                                                     __nv_bfloat16*       norm_output,
+                                                     const __nv_bfloat16* input,
+                                                     const __nv_bfloat16* gamma,
+                                                     const float          layernorm_eps,
+                                                     int                  m,
+                                                     int                  n,
+                                                     cudaStream_t         stream);
+#endif
+
 template<typename T>
 void invokeGeneralAddBiasResidualT5PreLayerNorm(T*           output,
                                                 T*           norm_output,