--- a/src/fastertransformer/models/llama/Llama.cc
+++ b/src/fastertransformer/models/llama/Llama.cc
@@ -1206,4 +1206,8 @@ bool* Llama<T>::getFinishBuffer()
 template class Llama<float>;
 template class Llama<half>;
 
+#ifdef ENABLE_BF16
+template class Llama<__nv_bfloat16>;
+#endif
+
 }  // namespace fastertransformer