--- a/src/fastertransformer/models/llama/Llama.cc
+++ b/src/fastertransformer/models/llama/Llama.cc
@@ -269,7 +269,11 @@ Llama<T>::Llama(size_t                              head_num,
     pipeline_para_.rank_       = 0;
 
     int local_vacab_size = ceil(vocab_size_ / 1.f / tensor_para_.world_size_);
-    if (std::is_same<half, T>::value) {
+    if (std::is_same<half, T>::value
+#ifdef ENABLE_BF16
+        || std::is_same<__nv_bfloat16, T>::value
+#endif
+    ) {
         local_vacab_size = ceil(local_vacab_size / 8.f) * 8;
     }
     vocab_size_padded_ = (size_t)local_vacab_size * tensor_para_.world_size_;