--- a/src/fastertransformer/models/llama/LlamaDecoder.cc
+++ b/src/fastertransformer/models/llama/LlamaDecoder.cc
@@ -377,8 +377,4 @@ void LlamaDecoder<T>::forward(std::unordered_map<std::string, Tensor>*          
 template class LlamaDecoder<float>;
 template class LlamaDecoder<half>;
 
-#ifdef ENABLE_BF16
-template class LlamaDecoder<__nv_bfloat16>;
-#endif
-
 }  // namespace fastertransformer