… CMakeLists.txt
… examples/cpp/CMakeLists.txt
… examples/cpp/llama/CMakeLists.txt
… examples/cpp/llama/bad_words.csv
… examples/cpp/llama/check_with_huggingface.py
… examples/cpp/llama/huggingface_llama_convert.py
… examples/cpp/llama/llama_config.ini
… examples/cpp/llama/llama_example.cc
… examples/cpp/llama/llama_triton_example.cc
… examples/cpp/llama/model_config.json
… examples/cpp/llama/start_ids.csv
… examples/cpp/llama/stop_words.csv
… examples/pytorch/gpt/utils/megatron_ckpt_convert_llama.py
… examples/pytorch/llama/llama_config.ini
… examples/pytorch/llama/llama_example.py
… examples/pytorch/llama/start_ids.csv
… examples/pytorch/llama/utils/llama.py
… src/fastertransformer/models/CMakeLists.txt
… src/fastertransformer/models/llama/CMakeLists.txt
… src/fastertransformer/models/llama/LlamaContextDecoder.h
… src/fastertransformer/models/llama/LlamaDecoder.h
… src/fastertransformer/models/llama/LlamaDecoderLayerWeight.h
… src/fastertransformer/models/llama/LlamaWeight.h
… src/fastertransformer/th_op/CMakeLists.txt
… src/fastertransformer/th_op/llama/CMakeLists.txt
… src/fastertransformer/th_op/llama/LlamaOp.cc
… src/fastertransformer/th_op/llama/LlamaOp.h
… src/fastertransformer/triton_backend/CMakeLists.txt
… src/fastertransformer/triton_backend/llama/CMakeLists.txt
… src/fastertransformer/triton_backend/llama/LlamaTritonModel.cc