Added pytorch backend support for llama, based on https://github.com/NVIDIA/FasterTransformer/pull/575.

For a simple test, run under `examples/pytorch/llama/`:
```bash
mpirun -n 4 --allow-run-as-root python llama_example.py --tensor_para_size=4 --pipeline_para_size=1  --ckpt_path path_to_weight --tokenizer_path path_to_tokenizer --lib_path path_to_libth_transformer.so --max_batch_size 1 --start_id_file start_ids.csv
```
It should print:
```
[Context]
0, 18637, 29892, 526, 366, 1136, 455, 2470, 29973, 1815, 366, 5193, 304, 592, 29973

[Generated]
[18637 29892   526   366  1136   455  2470 29973  1815   366  5193   304
   592 29973 18637 29892   526   366  1136   455  2470 29973  1815   366
  5193   304   592 29973 18637 29892   526   366  1136   455  2470 29973
  1815   366  5193   304   592 29973 18637 29892   526   366  1136]

[Output]
Hey, are you consciours? Can you talk to me? Hey, are you consciours? Can you talk to me? Hey, are you consciours? Can you talk to me? Hey, are you cons
```

which meets with the result of its cpp version.