@RomaA2000 2023-05-17
Is there any guide how to create correct gemm_config.in for llama parameters?