@hepj987 2023-05-18
Hello, may I ask if this change can support the CPP version of llama?Or does it only support the Pytorch version of llama?